v0.32a - transaction initial support

+++ app/src/main/java/com/niku/moneymate/account/Account.kt
@@ -16,7 +16,7 @@ import java.util.*
 data class Account(
     var currency_id: UUID,
     var title: String = "",
-    var balance: Int = 0,
+    var balance: Double = 0.0,
     var note: String = "",
     @PrimaryKey val account_id: UUID = UUID.randomUUID()
 )
